WIP fix tests
